+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### Version 1.51
+
+- Detect onesignal-cordova-plugin and @ionic-enterprise/auth conflict
+- Recommend Cocoapods update to 1.13.0 to fix issues with building with XCode 15
+
 ### Version 1.50
 
 - New Project now has Capacitor Plugin as an option